backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
openai
python-dotenv
pydantic
